raspi2 porting

+++ lib/src/kstring.c
@@ -1,14 +1,14 @@
-#include <string.h>
+#include "kstring.h"
 
 /*
  * memcmp copies n bytes from the source buffer to the target buffer. It returns
- * the pointer to the target.
+ * the point32_ter to the target.
  */
-void *memcpy(void *target, const void *source, size_t n)
+void *memcpy(void *target, const void *source, uint32_t n)
 {
 	char *target_buffer = (char *) target;
 	char *source_buffer = (char *) source;
-	size_t i = 0;
+	uint32_t i = 0;
 
 	for (i = 0; i < n; i++)
 		target_buffer[i] = source_buffer[i];
@@ -17,8 +17,8 @@ void *memcpy(void *target, const void *source, size_t n)
 }
 
 /*
- * strcpy copies the given null-terminated source string into the given target.
- * It returns the pointer to the target.
+ * strcpy copies the given null-terminated source string int32_to the given target.
+ * It returns the point32_ter to the target.
  */
 char *strcpy(char *target, const char *source)
 {
@@ -41,10 +41,10 @@ char *strcpy(char *target, const char *source)
  * function returns strlen(source). If return value is >= n, then truncation
  * occured.
  */
-size_t strncpy(char *target, const char *source, size_t n)
+uint32_t strncpy(char *target, const char *source, uint32_t n)
 {
-	size_t source_len = 0;
-	size_t i = 0;
+	uint32_t source_len = 0;
+	uint32_t i = 0;
 
 	while (i < n && source[i] != '\0') {
 		target[i] = source[i];
@@ -62,7 +62,7 @@ size_t strncpy(char *target, const char *source, size_t n)
  * returns 0 if s1 == s2, returns a negative number if s1 < s2, and returns a 
  * positive number if s1 > s2.
  */
-int strcmp(const char *s1, const char *s2)
+int32_t strcmp(const char *s1, const char *s2)
 {
 	while (*s1 == *s2 && *s1 != '\0' && *s2 != '\0') {
 		s1++;
@@ -77,9 +77,9 @@ int strcmp(const char *s1, const char *s2)
  * strncmp compares the given input strings upto n characters using the
  * lexicographical order. For return value, please see strcmp.
  */
-int strncmp(const char *s1, const char *s2, size_t n)
+int32_t strncmp(const char *s1, const char *s2, uint32_t n)
 {
-	size_t i = 0;
+	uint32_t i = 0;
 
 	if (n == 0)
 		return 0;
@@ -94,10 +94,10 @@ int strncmp(const char *s1, const char *s2, size_t n)
 }
 
 /*
- * strchr returns a pointer to the first occurence of the given character in the
+ * strchr returns a point32_ter to the first occurence of the given character in the
  * given string. If the character is not found, it returns NULL.
  */
-char *strchr(const char *str, int character)
+char *strchr(const char *str, int32_t character)
 {
 	while (*str != '\0' && *str != character)
 		str++;
@@ -110,8 +110,8 @@ char *strchr(const char *str, int character)
 
 /*
  * strtok tokenizes the given string using the given delimiters. If str != NULL,
- * then it returns a pointer to the first token. If str == NULL, then it returns
- * the a pointer to the next token of the string used in previous calls. If no
+ * then it returns a point32_ter to the first token. If str == NULL, then it returns
+ * the a point32_ter to the next token of the string used in previous calls. If no
  * more tokens are found, it returns NULL.
  *
  * WARNING: This function changes the original string.
@@ -150,11 +150,11 @@ char *strtok(char *str, const char *delimiters)
 }
 
 /* memset fills the given target with given length with the given character. */
-void *memset(void *target, int c, size_t len)
+void *memset(void *target, int32_t c, uint32_t len)
 {
 	char *target_buffer = (char *) target;
 
-	size_t i = 0;
+	uint32_t i = 0;
 	for (i = 0; i < len; i++)
 		target_buffer[i] = (char) c;
 
@@ -162,9 +162,9 @@ void *memset(void *target, int c, size_t len)
 }
 
 /* strlen returns the length of the given null-terminated string. */
-size_t strlen(const char *str)
+uint32_t strlen(const char *str)
 {
-	size_t length = 0;
+	uint32_t length = 0;
 	while (*str != '\0') {
 		str++;
 		length++;
