+++ system/lib/include/stdlib.h
@@ -3,7 +3,7 @@
 
 #include <types.h>
 
-void* malloc(size_t size);
+void* malloc(uint32_t size);
 
 void free(void* p);
