+++ system/lib/src/unistd.c
@@ -83,7 +83,7 @@ void yield() {
 	syscall0(SYSCALL_YIELD);
 }
 
-char* getcwd(char* buf, size_t size) {
+char* getcwd(char* buf, uint32_t size) {
 	return (char*)syscall2(SYSCALL_GET_CWD, (int)buf, (int)size);
 }
